FRF.16: Add a test capture file
This capture file triggered a buffer over-read with the 4.9.2 release
and was fixed in the 4.9.3 release.

With the new GET_ macros, the problem no longer exists.
Thus no fix needed.

The pcap file was updated to pass the sanity checks and to avoid
pcap_loop() error messages.

Update the output of the test accordingly.

Partial update from aa3e54f594385ce7e1e319b0c84999e51192578b
in 4.9 branch.